Move files into matching folders.
Webpack also moves scss files into dist. Will share the same folder structure for scss and js, so that header component scss and js sit side by side.
