+++ scss/components/header/_header.scss
@@ -1,7 +1,7 @@
-@use "../settings/breakpoints" as *;
-@use "../tools/colour";
-@use "../tools/spacing";
-@use "../tools/typography";
+@use "../../settings/breakpoints" as *;
+@use "../../tools/colour";
+@use "../../tools/spacing";
+@use "../../tools/typography";
 
 /* Site Header Styles */
