+++ scss/components/index.scss
@@ -1,5 +1,5 @@
 @use "footer";
-@use "header";
+@use "header/header";
 @use "link";
 @use "lists";
 @use "logo";
