+++ src/tools/getOrThrowIfNull.spec.ts
